+++ Editor/AnimationRig/Constraints/MultiAimConstraintEditor.cs
@@ -9,15 +9,45 @@ namespace UnityEditor.Animations.Rigging
     [CanEditMultipleObjects]
     class MultiAimConstraintEditor : Editor
     {
-        static readonly GUIContent k_SourceObjectsLabel = new GUIContent("Source Objects");
-        static readonly GUIContent k_SettingsLabel = new GUIContent("Settings");
-        static readonly GUIContent k_AimAxisLabel = new GUIContent("Aim Axis");
-        static readonly GUIContent k_UpAxisLabel = new GUIContent("Up Axis");
-        static readonly GUIContent k_WorldUpAxisLabel = new GUIContent("World Up Axis");
-        static readonly GUIContent k_WorldUpType = new GUIContent("World Up Type");
-        static readonly GUIContent k_WorldUpObject = new GUIContent("World Up Object");
-        static readonly GUIContent[] k_AxisLabels = new []{ "X", "-X", "Y", "-Y", "Z", "-Z" }.Select(s => new GUIContent(s)).ToArray();
-        static readonly GUIContent k_MaintainOffsetLabel = new GUIContent("Maintain Rotation Offset");
+        static class Content
+        {
+            public static readonly GUIContent[] axisLabels = new []{ "X", "-X", "Y", "-Y", "Z", "-Z" }
+                .Select(c => new GUIContent(c))
+                .ToArray();
+            public static readonly GUIContent aimAxis = EditorGUIUtility.TrTextContent(
+                "Aim Axis",
+                "Specifies the local aim axis of the Constrained Object to use in order to orient its forward direction to the Source Objects."
+            );
+            public static readonly GUIContent upAxis = EditorGUIUtility.TrTextContent(
+                "Up Axis",
+                "Specifies the local up axis of the Constrained Object to use in order to orient its upward direction (i.e., roll orientation)."
+            );
+            public static readonly GUIContent worldUpType = EditorGUIUtility.TrTextContent(
+                "World Up Type",
+                "Specifies which mode to use to stabilize the upward direction (i.e., roll orientation) of the Constrained Object."
+            );
+            public static readonly GUIContent worldUpAxis = EditorGUIUtility.TrTextContent(
+                "World Up Axis",
+                "A vector in some reference frame that is used to stabilize the upward direction of the Constrained Object. " +
+                "This value is used when World Up Type is either Vector or Object Rotation Up."
+            );
+            public static readonly GUIContent worldUpObject = EditorGUIUtility.TrTextContent(
+                "World Up Object",
+                "A GameObject used as a reference frame for World Up Axis. " +
+                "This value is used when World Up Type is either Object Up or Object Rotation Up."
+            );
+            public static readonly GUIContent sourceObjects = CommonContent.sourceObjectsWeightedRotation;
+            public static readonly GUIContent settings = CommonContent.settings;
+            public static readonly GUIContent maintainOffset = CommonContent.maintainRotationOffset;
+            public static readonly GUIContent minLimit = EditorGUIUtility.TrTextContent(
+                "Min Limit",
+                "Clamps the minimum rotation that may be applied about any of the constrained axes of rotation."
+            );
+            public static readonly GUIContent maxLimit = EditorGUIUtility.TrTextContent(
+                "Max Limit",
+                "Clamps the maximum rotation that may be applied about any of the constrained axes of rotation."
+            );
+        }
 
         SerializedProperty m_Weight;
         SerializedProperty m_ConstrainedObject;
@@ -58,41 +88,41 @@ namespace UnityEditor.Animations.Rigging
         {
             serializedObject.Update();
 
-            EditorGUILayout.PropertyField(m_Weight);
+            EditorGUILayout.PropertyField(m_Weight, CommonContent.weight);
 
-            EditorGUILayout.PropertyField(m_ConstrainedObject);
+            EditorGUILayout.PropertyField(m_ConstrainedObject, CommonContent.constrainedObject);
 
             ++EditorGUI.indentLevel;
-            DoAxisField(m_AimAxis, k_AimAxisLabel);
-            DoAxisField(m_UpAxis, k_UpAxisLabel);
+            DoAxisField(m_AimAxis, Content.aimAxis);
+            DoAxisField(m_UpAxis, Content.upAxis);
             --EditorGUI.indentLevel;
 
-            EditorGUILayout.PropertyField(m_WorldUpType, k_WorldUpType);
+            EditorGUILayout.PropertyField(m_WorldUpType, Content.worldUpType);
 
             var worldUpType = (MultiAimConstraintData.WorldUpType)m_WorldUpType.intValue;
 
             ++EditorGUI.indentLevel;
             using (new EditorGUI.DisabledGroupScope(worldUpType != MultiAimConstraintData.WorldUpType.ObjectRotationUp && worldUpType != MultiAimConstraintData.WorldUpType.Vector))
             {
-                DoAxisField(m_WorldUpAxis, k_WorldUpAxisLabel);
+                DoAxisField(m_WorldUpAxis, Content.worldUpAxis);
             }
             using (new EditorGUI.DisabledGroupScope(worldUpType != MultiAimConstraintData.WorldUpType.ObjectUp && worldUpType != MultiAimConstraintData.WorldUpType.ObjectRotationUp))
             {
-                EditorGUILayout.PropertyField(m_WorldUpObject, k_WorldUpObject);
+                EditorGUILayout.PropertyField(m_WorldUpObject, Content.worldUpObject);
             }
             --EditorGUI.indentLevel;
 
-            EditorGUILayout.PropertyField(m_SourceObjects, k_SourceObjectsLabel);
+            EditorGUILayout.PropertyField(m_SourceObjects, Content.sourceObjects);
 
-            m_SettingsToggle.value = EditorGUILayout.BeginFoldoutHeaderGroup(m_SettingsToggle.value, k_SettingsLabel);
+            m_SettingsToggle.value = EditorGUILayout.BeginFoldoutHeaderGroup(m_SettingsToggle.value, Content.settings);
             if (m_SettingsToggle.value)
             {
                 EditorGUI.indentLevel++;
-                EditorGUILayout.PropertyField(m_MaintainOffset, k_MaintainOffsetLabel);
-                EditorGUILayout.PropertyField(m_Offset);
-                EditorGUILayout.PropertyField(m_ConstrainedAxes);
-                EditorGUILayout.PropertyField(m_MinLimit);
-                EditorGUILayout.PropertyField(m_MaxLimit);
+                EditorGUILayout.PropertyField(m_MaintainOffset, Content.maintainOffset);
+                EditorGUILayout.PropertyField(m_Offset, CommonContent.offsetRotation);
+                EditorGUILayout.PropertyField(m_ConstrainedAxes, CommonContent.constrainedAxesRotation);
+                EditorGUILayout.PropertyField(m_MinLimit, Content.minLimit);
+                EditorGUILayout.PropertyField(m_MaxLimit, Content.maxLimit);
                 EditorGUI.indentLevel--;
             }
             EditorGUILayout.EndFoldoutHeaderGroup();
@@ -105,7 +135,7 @@ namespace UnityEditor.Animations.Rigging
             var rect = EditorGUILayout.GetControlRect();
             EditorGUI.BeginProperty(rect, label, property);
             EditorGUI.BeginChangeCheck();
-            var newValue = EditorGUI.Popup(rect, label, property.intValue, k_AxisLabels);
+            var newValue = EditorGUI.Popup(rect, label, property.intValue, Content.axisLabels);
             if (EditorGUI.EndChangeCheck())
                 property.intValue = newValue;
             EditorGUI.EndProperty();